Adding Unit Tests for Contested TemplateNames
